requirements.txt
Appium-Python-Client==1.0.2
atomicwrites==1.4.0
attrs==19.3.0
colorama==0.4.3
iniconfig==1.0.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pyparsing==2.4.7
pytest==6.0.1
pytest-sugar==0.9.4
selenium==3.141.0
six==1.15.0
termcolor==1.1.0
toml==0.10.1
urllib3==1.25.10
